feat: Allow region overrides in service path
This adds a new lookup for regional overrides in services/${service}/region_overrides/${region}/${cluster_name}.yaml. The intent is for this to ultimately replace the clusters/${cluster}/default.yaml service overrides so that they are more clearly broken out and easier to find. Also allows a more natural splitting out of service files rather than using the include workaround we're currently using.

The current order of variables getting loaded is:

1. services/${service}_values.yaml
2. services/${service}/region_overrides/${region}/${cluster_name}.yaml
3. clusters/${cluster}/${cluster_name}.yaml

This is a backport of https://github.com/getsentry/ops/pull/11548