fix(service_registry): Dynamic detection of path
Since the service registry is moving to its own repository, we need
a mechanism to get the relevant data from there. This PR introduces
a mechanism which first checks whether the service registry package
is installed or not. If yes, uses the file from the package. If not,
falls back to hardcoded path from before.