Use own jsonpatch implementation (#55)
This PR deprecates the use of jsonpatch in quickpatch and instead uses
our own patcher:

The patcher interprets the paths defined in the .managed.yaml file
according to a few rules:
1. Paths must be separated by `/`
2. Paths must refer to nested json objs in the corresponding values file
that is being patched
3. If the json obj in the `.managed.yaml` file does not exist, the
patcher will auto-generate all of the relevant objs as needed & defined
by the patch
4. If the values file does not exist, the patcher will also
auto-generate the `.managed.yaml` file
5. Any values that are defined in the patch will be set at the path
location. Aka, the patcher may override existing objs at the defined
path. To avoid this issue, specify the full length of the path down to
the lowest-nested level, with the value containing a single scalar or
string. See the added `test-patch-override.yaml` file for example of
this override happening.