Prevents --context from falling after the "--". (#167)
Moving context to the end fails with commands that construct an arg list
with "--" in it (including run-pod and exec). "--" indicates the end of
arguments, and causes the context flag to not be accepted by kubectl.
This PR checks if there is a "--" and if so, puts the context flag
immediately before it.