feat: add include_raw jinja global (#168)
This PR adds an `include_raw()` Jinja global, which works the same as
`{% include 'path/to/file.yaml' %}` but doesn't render any Jinja in that
file.

The code in this PR is basically an updated version of
https://stackoverflow.com/questions/9767585/insert-static-files-literally-into-jinja-templates-without-parsing-them.

This is needed for the workflow engine Hera API because, when
materializing a workflow that uses a `ScriptTemplate`, Hera inserts a
block like this at the top of the script in the materialized YAML file:
```
import os
import sys
sys.path.append(os.getcwd())
import json
try: customer_name = json.loads(r'''{{inputs.parameters.customer_name}}''')
except: customer_name = r'''{{inputs.parameters.customer_name}}'''
```

The `{{inputs.parameters.customer_name}}` jinja vars are meant to be
used by Argo, but `sentry-kube` attempts to render the jinja then
crashes because it doesn't know what `inputs` is referring to.

There's no way to insert `{% raw %}` blocks before the above lines in
the materialized YAML file, as the entire materialized file is created
by Hera from Python code. Therefore I had to make a change to
`sentry-kube`.

---------

Co-authored-by: Richard Gibert <richard@gibert.ca>